updating eventbus
