+++ crates/opsml_events/src/lib.rs
@@ -1,2 +1,5 @@
 pub mod event;
 pub mod types;
+
+pub use event::EventBus;
+pub use types::{AuditEvent, Event};
