+++ crates/opsml_server/src/core/experiment/route.rs
@@ -1,10 +1,6 @@
-use crate::core::audit::{spawn_audit_task, AuditArgs};
 use crate::core::experiment::types::GroupedMetric;
 /// Route for checking if a card UID exists
-use crate::core::{
-    error::{internal_server_error, internal_server_error_with_audit},
-    state::AppState,
-};
+use crate::core::{error::internal_server_error, state::AppState};
 use anyhow::{Context, Result};
 use axum::{
     extract::{ConnectInfo, Query, State},
@@ -15,6 +11,7 @@ use axum::{
 use axum_extra::TypedHeader;
 use headers::UserAgent;
 use opsml_client::Routes;
+use opsml_events::{AuditEvent, Event};
 use opsml_sql::base::SqlClient;
 use opsml_sql::schemas::schema::{HardwareMetricsRecord, MetricRecord, ParameterRecord};
 use opsml_types::{cards::*, contracts::*, RegistryType};
@@ -35,7 +32,7 @@ pub async fn insert_metrics(
     headers: HeaderMap,
     Json(req): Json<MetricRequest>,
 ) -> Result<Json<MetricResponse>, (StatusCode, Json<serde_json::Value>)> {
-    let audit_args = AuditArgs::new(
+    let audit_args = AuditEvent::new(
         addr,
         agent,
         headers,
@@ -66,16 +63,11 @@ pub async fn insert_metrics(
         .await
         .map_err(|e| {
             error!("Failed to insert metric: {}", e);
-            internal_server_error_with_audit(
-                e,
-                "Failed to insert metric",
-                audit_args.clone(),
-                state.sql_client.clone(),
-            )
+            internal_server_error(e, "Failed to insert metric")
         })?;
 
     // Spawn audit task
-    spawn_audit_task(audit_args, state.sql_client.clone());
+    //spawn_audit_task(audit_args, state.sql_client.clone());
     Ok(Json(MetricResponse { success: true }))
 }
 
@@ -86,7 +78,7 @@ pub async fn get_metrics(
     headers: HeaderMap,
     Json(req): Json<GetMetricRequest>,
 ) -> Result<Json<Vec<Metric>>, (StatusCode, Json<serde_json::Value>)> {
-    let audit_args = AuditArgs::new(
+    let audit_args = AuditEvent::new(
         addr,
         agent,
         headers,
@@ -103,12 +95,7 @@ pub async fn get_metrics(
         .await
         .map_err(|e| {
             error!("Failed to get metrics: {}", e);
-            internal_server_error_with_audit(
-                e,
-                "Failed to get metrics",
-                audit_args.clone(),
-                state.sql_client.clone(),
-            )
+            internal_server_error(e, "Failed to get metrics")
         })?;
 
     // map all entries in the metrics to the Metric struct
@@ -123,7 +110,7 @@ pub async fn get_metrics(
         })
         .collect::<Vec<_>>();
 
-    spawn_audit_task(audit_args, state.sql_client.clone());
+    // spawn_audit_task(audit_args, state.sql_client.clone());
     Ok(Json(metrics))
 }
 
@@ -136,7 +123,7 @@ pub async fn get_grouped_metrics(
 ) -> Result<Json<HashMap<String, Vec<GroupedMetric>>>, (StatusCode, Json<serde_json::Value>)> {
     let mut metric_data: HashMap<String, Vec<GroupedMetric>> = HashMap::new();
 
-    let audit_args = AuditArgs::new(
+    let audit_args = AuditEvent::new(
         addr,
         agent,
         headers,
@@ -155,12 +142,7 @@ pub async fn get_grouped_metrics(
             .await
             .map_err(|e| {
                 error!("Failed to get metrics: {}", e);
-                internal_server_error_with_audit(
-                    e,
-                    "Failed to get metrics",
-                    audit_args.clone(),
-                    state.sql_client.clone(),
-                )
+                internal_server_error(e, "Failed to get metrics")
             })?;
 
         let mut grouped_by_name: HashMap<String, Vec<MetricRecord>> = HashMap::new();
@@ -201,7 +183,7 @@ pub async fn get_grouped_metrics(
         }
     }
 
-    spawn_audit_task(audit_args, state.sql_client.clone());
+    //spawn_audit_task(audit_args, state.sql_client.clone());
     Ok(Json(metric_data))
 }
 
@@ -212,7 +194,7 @@ pub async fn get_metric_names(
     headers: HeaderMap,
     Query(req): Query<GetMetricNamesRequest>,
 ) -> Result<Json<Vec<String>>, (StatusCode, Json<serde_json::Value>)> {
-    let audit_args = AuditArgs::new(
+    let audit_args = AuditEvent::new(
         addr,
         agent,
         headers,
@@ -229,15 +211,10 @@ pub async fn get_metric_names(
         .await
         .map_err(|e| {
             error!("Failed to get metrics: {}", e);
-            internal_server_error_with_audit(
-                e,
-                "Failed to get metrics",
-                audit_args.clone(),
-                state.sql_client.clone(),
-            )
+            internal_server_error(e, "Failed to get metrics")
         })?;
 
-    spawn_audit_task(audit_args, state.sql_client.clone());
+    //spawn_audit_task(audit_args, state.sql_client.clone());
     Ok(Json(names))
 }
 
@@ -248,7 +225,7 @@ pub async fn insert_parameters(
     headers: HeaderMap,
     Json(req): Json<ParameterRequest>,
 ) -> Result<Json<ParameterResponse>, (StatusCode, Json<serde_json::Value>)> {
-    let audit_args = AuditArgs::new(
+    let audit_args = AuditEvent::new(
         addr,
         agent,
         headers,
@@ -271,16 +248,11 @@ pub async fn insert_parameters(
         .await
         .map_err(|e| {
             error!("Failed to insert parameter: {}", e);
-            internal_server_error_with_audit(
-                e,
-                "Failed to insert parameter",
-                audit_args.clone(),
-                state.sql_client.clone(),
-            )
+            internal_server_error(e, "Failed to insert parameter")
         })?;
 
     // Spawn audit task
-    spawn_audit_task(audit_args, state.sql_client.clone());
+    //spawn_audit_task(audit_args, state.sql_client.clone());
     Ok(Json(ParameterResponse { success: true }))
 }
 
@@ -291,7 +263,7 @@ pub async fn get_parameter(
     headers: HeaderMap,
     Json(req): Json<GetParameterRequest>,
 ) -> Result<Json<Vec<Parameter>>, (StatusCode, Json<serde_json::Value>)> {
-    let audit_args = AuditArgs::new(
+    let audit_args = AuditEvent::new(
         addr,
         agent,
         headers,
@@ -308,12 +280,7 @@ pub async fn get_parameter(
         .await
         .map_err(|e| {
             error!("Failed to get metrics: {}", e);
-            internal_server_error_with_audit(
-                e,
-                "Failed to get metrics",
-                audit_args.clone(),
-                state.sql_client.clone(),
-            )
+            internal_server_error(e, "Failed to get metrics")
         })?;
 
     // map all entries in the metrics to the Metric struct
@@ -325,7 +292,7 @@ pub async fn get_parameter(
         })
         .collect::<Vec<_>>();
 
-    spawn_audit_task(audit_args, state.sql_client.clone());
+    //spawn_audit_task(audit_args, state.sql_client.clone());
     Ok(Json(params))
 }
 
@@ -336,6 +303,17 @@ pub async fn insert_hardware_metrics(
     headers: HeaderMap,
     Json(req): Json<HardwareMetricRequest>,
 ) -> Result<Json<HardwareMetricResponse>, (StatusCode, Json<serde_json::Value>)> {
+    let audit_args = AuditEvent::new(
+        addr,
+        agent,
+        headers,
+        Operation::Write,
+        ResourceType::Database,
+        req.experiment_uid.clone(),
+        serde_json::to_string(&req).unwrap_or_default(),
+        RegistryType::Experiment,
+        Routes::ExperimentHardwareMetrics,
+    );
     let created_at = get_utc_datetime();
 
     let record = HardwareMetricsRecord {
