+++ crates/opsml_server/src/core/state.rs
@@ -2,6 +2,7 @@ use crate::core::scouter::client::ScouterApiClient;
 use opsml_auth::auth::AuthManager;
 use opsml_auth::permission::UserPermissions;
 use opsml_error::ApiError;
+use opsml_events::EventBus;
 use opsml_settings::config::{OpsmlConfig, OpsmlStorageSettings};
 use opsml_sql::base::SqlClient;
 use opsml_sql::enums::client::SqlClientEnum;
@@ -16,6 +17,7 @@ pub struct AppState {
     pub config: Arc<OpsmlConfig>,
     pub storage_settings: Arc<OpsmlStorageSettings>,
     pub scouter_client: Arc<ScouterApiClient>,
+    pub event_bus: EventBus,
 }
 
 impl AppState {
